Fixed lagged times
Lagged times appear when the client's ping fluctuates throughout a run.
For example, if I start a run with a ping of 500 ms the starttimer
trigger will arrive at the server half a second after I actually
started. If I then finish after 4 seconds with a ping of 0, sending the
stoptimer immediately, the recorded time will be 3.5 seconds. This may
happen frequently when someone in your network is downloading
some small files while you are playing a small map.

To compensate this, the client's ping is now subtracted whenever a time
is saved.

Signed-off-by: Gerco van Heerdt <gercovanheerdt@gmail.com>